fix(security): resolve CodeQL alerts for unused variables and race conditions
- Remove unused variable 'tableInfo' in check-holidays-policies.js
- Remove unused import 'path' in scripts/fix-console-logs.js
- Replace problematic scripts with secure versions that prevent race conditions
- Add file change verification before writing to prevent race conditions
- Delete old scripts that had security vulnerabilities
- Maintain all functionality while improving security
