+++ src/lib/admin-query.ts
@@ -2,8 +2,8 @@
 
 import { supabaseAdmin } from '@/lib/supabase-admin';
 import type { AdminInsert, User as AdminUser } from '@/types';
-
 import { securityLogger } from '@/utils/security-config';
+
 /**
  * Crea un nuevo usuario administrador en Supabase.
  * Esta función debe ser llamada desde un entorno seguro o por un usuario con permisos de super_admin.
@@ -266,7 +266,7 @@ export const getUsersStats = async (): Promise<{
       .select('id, role');
 
     if (error !== null) {
-            securityLogger.error('Error fetching users stats:', error);
+      securityLogger.error('Error fetching users stats:', error);
       throw error;
     }
 
@@ -282,7 +282,7 @@ export const getUsersStats = async (): Promise<{
 
     return stats;
   } catch (error) {
-        securityLogger.error('Error in getUsersStats:', error);
+    securityLogger.error('Error in getUsersStats:', error);
     throw error;
   }
 };
