+++ src/lib/workers-query.ts
@@ -1,6 +1,6 @@
 import type { Worker, WorkerInsert } from '@/types';
-
 import { securityLogger } from '@/utils/security-config';
+
 import { supabase } from './database';
 
 /**
@@ -14,13 +14,13 @@ export const getAllWorkers = async (): Promise<Worker[]> => {
       .order('created_at', { ascending: false });
 
     if (error !== null) {
-            securityLogger.error('Error fetching workers:', error);
+      securityLogger.error('Error fetching workers:', error);
       throw error;
     }
 
     return (data ?? []) as Worker[];
   } catch (error) {
-        securityLogger.error('Error in getAllWorkers:', error);
+    securityLogger.error('Error in getAllWorkers:', error);
     throw error;
   }
 };
@@ -37,13 +37,13 @@ export const getActiveWorkers = async (): Promise<Worker[]> => {
       .order('name');
 
     if (error !== null) {
-            securityLogger.error('Error fetching active workers:', error);
+      securityLogger.error('Error fetching active workers:', error);
       throw error;
     }
 
     return (data ?? []) as Worker[];
   } catch (error) {
-        securityLogger.error('Error in getActiveWorkers:', error);
+    securityLogger.error('Error in getActiveWorkers:', error);
     throw error;
   }
 };
@@ -63,13 +63,13 @@ export const getWorkersByType = async (
       .order('name');
 
     if (error !== null) {
-            securityLogger.error('Error fetching workers by type:', error);
+      securityLogger.error('Error fetching workers by type:', error);
       throw error;
     }
 
     return (data ?? []) as Worker[];
   } catch (error) {
-        securityLogger.error('Error in getWorkersByType:', error);
+    securityLogger.error('Error in getWorkersByType:', error);
     throw error;
   }
 };
@@ -93,7 +93,7 @@ export const getWorkersStats = async (): Promise<WorkersStats> => {
       .select('worker_type, is_active');
 
     if (error !== null) {
-            securityLogger.error('Error fetching workers stats:', error);
+      securityLogger.error('Error fetching workers stats:', error);
       throw error;
     }
 
@@ -114,7 +114,7 @@ export const getWorkersStats = async (): Promise<WorkersStats> => {
 
     return stats;
   } catch (error) {
-        securityLogger.error('Error in getWorkersStats:', error);
+    securityLogger.error('Error in getWorkersStats:', error);
     throw error;
   }
 };
@@ -134,13 +134,13 @@ export const searchWorkers = async (searchTerm: string): Promise<Worker[]> => {
       .order('name');
 
     if (error !== null) {
-            securityLogger.error('Error searching workers:', error);
+      securityLogger.error('Error searching workers:', error);
       throw error;
     }
 
     return (data ?? []) as Worker[];
   } catch (error) {
-        securityLogger.error('Error in searchWorkers:', error);
+    securityLogger.error('Error in searchWorkers:', error);
     throw error;
   }
 };
@@ -157,13 +157,13 @@ export const getWorkerById = async (id: string): Promise<Worker | null> => {
       .single();
 
     if (error !== null) {
-            securityLogger.error('Error fetching worker by ID:', error);
+      securityLogger.error('Error fetching worker by ID:', error);
       throw error;
     }
 
     return data as Worker;
   } catch (error) {
-        securityLogger.error('Error in getWorkerById:', error);
+    securityLogger.error('Error in getWorkerById:', error);
     throw error;
   }
 };
@@ -184,13 +184,13 @@ export const updateWorker = async (
       .single();
 
     if (error) {
-            securityLogger.error('Error updating worker:', error);
+      securityLogger.error('Error updating worker:', error);
       throw error;
     }
 
     return data as Worker;
   } catch (error) {
-        securityLogger.error('Error in updateWorker:', error);
+    securityLogger.error('Error in updateWorker:', error);
     throw error;
   }
 };
@@ -258,11 +258,11 @@ export const deleteWorker = async (id: string): Promise<void> => {
     const { error } = await supabase.from('workers').delete().eq('id', id);
 
     if (error) {
-            securityLogger.error('Error deleting worker:', error);
+      securityLogger.error('Error deleting worker:', error);
       throw error;
     }
   } catch (error) {
-        securityLogger.error('Error in deleteWorker:', error);
+    securityLogger.error('Error in deleteWorker:', error);
     throw error;
   }
 };
